move click actions, update imports
