+++ frontend/src/metabase/static-viz/components/RowChart/RowChart.tsx
@@ -11,8 +11,8 @@ import {
   trimData,
 } from "metabase/visualizations/shared/utils/data";
 import { getChartGoal } from "metabase/visualizations/lib/settings/goal";
-import { DatasetData, VisualizationSettings } from "metabase-types/api";
-import { ColorGetter } from "metabase/static-viz/lib/colors";
+import type { DatasetData, VisualizationSettings } from "metabase-types/api";
+import type { ColorGetter } from "metabase/static-viz/lib/colors";
 import { getTwoDimensionalChartSeries } from "metabase/visualizations/shared/utils/series";
 import {
   getAxesVisibility,
@@ -25,7 +25,7 @@ import {
   getStaticFormatters,
 } from "metabase/static-viz/lib/format";
 import { extractRemappedColumns } from "metabase/visualizations";
-import { RemappingHydratedChartData } from "metabase/visualizations/types";
+import type { RemappingHydratedChartData } from "metabase/visualizations/types";
 import { calculateLegendRows } from "../Legend/utils";
 import { Legend } from "../Legend";
