+++ frontend/src/metabase/visualizations/click-actions/actions/NativeQueryClickFallback.ts
@@ -1,10 +1,7 @@
 import { t } from "ttag";
 import { isWithinIframe } from "metabase/lib/dom";
 import { getEngineNativeType } from "metabase/lib/engine";
-import type {
-  RegularClickAction,
-  Drill,
-} from "metabase/visualizations/click-actions/types";
+import type { RegularClickAction, Drill } from "metabase/visualizations/types";
 import { nativeDrillFallback } from "metabase-lib/queries/drills/native-drill-fallback";
 
 export const NativeQueryClickFallback: Drill = ({ question }) => {
