+++ frontend/src/metabase/visualizations/click-actions/drills/ForeignKeyDrill.ts
@@ -1,6 +1,6 @@
 import { t } from "ttag";
 import { pluralize, singularize } from "metabase/lib/formatting/strings";
-import { Drill } from "metabase/visualizations/click-actions/types";
+import { Drill } from "metabase/visualizations/types";
 import {
   foreignKeyDrill,
   foreignKeyDrillQuestion,
