+++ frontend/src/metabase/visualizations/click-actions/drills/ObjectDetailDrill.ts
@@ -1,8 +1,8 @@
 import { t } from "ttag";
 import { zoomInRow } from "metabase/query_builder/actions";
 import type { DatasetColumn, RowValue } from "metabase-types/api";
-import type { ClickActionProps } from "metabase/visualizations/click-actions/types";
-import { DefaultClickAction } from "metabase/visualizations/click-actions/types";
+import type { ClickActionProps } from "metabase/visualizations/types";
+import { DefaultClickAction } from "metabase/visualizations/types";
 import type Question from "metabase-lib/Question";
 import {
   objectDetailDrill,
