+++ frontend/src/metabase/visualizations/click-actions/drills/ZoomDrill.unit.spec.ts
@@ -11,7 +11,7 @@ import {
   DatetimeUnit,
   StructuredDatasetQuery,
 } from "metabase-types/api";
-import type { ClickObject } from "metabase/visualizations/click-actions/types";
+import type { ClickObject } from "metabase/visualizations/types";
 import { checkNotNull } from "metabase/core/utils/types";
 import type Question from "metabase-lib/Question";
 import ZoomDrill from "./ZoomDrill";
