+++ frontend/src/metabase/visualizations/click-actions/modes/MetabotMode.ts
@@ -1,4 +1,4 @@
-import type { QueryMode } from "../types";
+import type { QueryMode } from "metabase/visualizations/types";
 
 export const MetabotMode: QueryMode = {
   name: "public",
