+++ frontend/src/metabase/visualizations/click-actions/modes/ObjectMode.ts
@@ -1,4 +1,4 @@
-import type { QueryMode } from "../types";
+import type { QueryMode } from "metabase/visualizations/types";
 import { ObjectDetailDrill } from "../drills/ObjectDetailDrill";
 
 export const ObjectMode: QueryMode = {
