+++ frontend/src/metabase/visualizations/click-actions/modes/PivotMode.ts
@@ -1,4 +1,4 @@
-import type { QueryMode } from "../types";
+import type { QueryMode } from "metabase/visualizations/types";
 import { PivotDrill } from "../drills/PivotDrill";
 import { DefaultMode } from "./DefaultMode";
