+++ frontend/src/metabase/visualizations/components/ChartTooltip/StackedDataTooltip/utils.ts
@@ -1,5 +1,5 @@
 import { t } from "ttag";
-import { TooltipRowModel } from "metabase/visualizations/types";
+import type { TooltipRowModel } from "metabase/visualizations/types";
 
 export const getTotalValue = (
   headerRows: TooltipRowModel[] = [],
