+++ frontend/src/metabase/visualizations/components/ChartTooltip/TimelineEventTooltip.tsx
@@ -1,6 +1,6 @@
 import { Icon, IconName } from "metabase/core/components/Icon";
 
-import { HoveredTimelineEvent } from "metabase/visualizations/types";
+import type { HoveredTimelineEvent } from "metabase/visualizations/types";
 import {
   TimelineEventDate,
   TimelineEventIconContainer,
