+++ frontend/src/metabase/visualizations/components/ClickActions/ClickActionControl.tsx
@@ -1,7 +1,5 @@
-import {
-  ClickAction,
-  isRegularClickAction,
-} from "metabase/visualizations/click-actions/types";
+import type { ClickAction } from "metabase/visualizations/types";
+import { isRegularClickAction } from "metabase/visualizations/types";
 import Tooltip from "metabase/core/components/Tooltip";
 import { color } from "metabase/lib/colors";
 import { Icon, IconName } from "metabase/core/components/Icon";
