+++ frontend/src/metabase/visualizations/components/ClickActions/ClickActionsPopover.tsx
@@ -4,16 +4,18 @@ import * as tippy from "tippy.js";
 import * as MetabaseAnalytics from "metabase/lib/analytics";
 import { getEventTarget } from "metabase/lib/dom";
 import { performAction } from "metabase/visualizations/lib/action";
-import { OnChangeCardAndRun } from "metabase/visualizations/types";
 import { Dispatch } from "metabase-types/store";
 import { Series } from "metabase-types/api";
-import {
+import type {
   RegularClickAction,
   ClickObject,
   PopoverClickAction,
+  OnChangeCardAndRun,
+} from "metabase/visualizations/types";
+import {
   isPopoverClickAction,
   isRegularClickAction,
-} from "metabase/visualizations/click-actions/types";
+} from "metabase/visualizations/types";
 
 import { ClickActionsView } from "./ClickActionsView";
 import { getGALabelForAction } from "./utils";
