+++ frontend/src/metabase/visualizations/lib/action.unit.spec.ts
@@ -1,4 +1,4 @@
-import { UrlClickAction } from "metabase/visualizations/click-actions/types";
+import type { UrlClickAction } from "metabase/visualizations/types";
 import { performAction } from "./action";
 
 describe("performAction", () => {
