+++ frontend/src/metabase/visualizations/types/click-actions.ts
