+++ frontend/src/metabase/visualizations/types/index.ts
@@ -1,3 +1,4 @@
 export * from "./hover";
 export * from "./columns";
 export * from "./visualization";
+export * from "./click-actions";
