+++ frontend/src/metabase/visualizations/visualizations/PieChart/utils.ts
@@ -1,5 +1,5 @@
 import _ from "underscore";
-import { StackedTooltipModel } from "metabase/visualizations/types";
+import type { StackedTooltipModel } from "metabase/visualizations/types";
 
 export function getMaxLabelDimension(
   d3Arc: d3.svg.Arc<d3.svg.arc.Arc>,
