+++ frontend/src/metabase/visualizations/visualizations/RowChart/utils/events.unit.spec.ts
@@ -2,7 +2,7 @@ import { SeriesSettings, VisualizationSettings } from "metabase-types/api";
 import { MultipleMetricsChartColumns } from "metabase/visualizations/lib/graph/columns";
 import { BarData } from "metabase/visualizations/shared/components/RowChart/types";
 import { GroupedDatum } from "metabase/visualizations/shared/types/data";
-import { RemappingHydratedDatasetColumn } from "metabase/visualizations/types";
+import type { RemappingHydratedDatasetColumn } from "metabase/visualizations/types";
 import { getHoverData } from "./events";
 
 const datasetColumns = [
